+++ data/my/src/main/java/com/hankki/data/my/repositoryimpl/MyRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.hankki.data.my.repositoryimpl
 
 import com.hankki.data.my.datasource.MyDataSource
-import com.hankki.domain.my.entity.MyRepository
+import com.hankki.domain.my.repository.MyRepository
 import javax.inject.Inject
 
 class MyRepositoryImpl @Inject constructor(
